Fix merkle code
